Add base class for summary implementations
