+++ src/main/java/dqm/jku/dqmeerkat/quality/profilingstatistics/singlecolumn/summary/SpaceSavingSummaryProfileStatistic.java
@@ -1,0 +1,32 @@
+package dqm.jku.dqmeerkat.quality.profilingstatistics.singlecolumn.summary;
+
+import dqm.jku.dqmeerkat.quality.DataProfile;
+import dqm.jku.dqmeerkat.quality.profilingstatistics.ProfileStatistic;
+import dqm.jku.dqmeerkat.quality.profilingstatistics.StatisticCategory;
+import dqm.jku.dqmeerkat.quality.profilingstatistics.StatisticTitle;
+
+import java.util.List;
+
+/**
+ * <h2>SpaceSavingSummaryProfileStatistic</h2>
+ * <summary>TODO Insert do cheader</summary>
+ *
+ * @author meindl, rainer.meindl@scch.at
+ * @since 06.07.2022
+ */
+public class SpaceSavingSummaryProfileStatistic extends SummaryProfileStatistic {
+
+    public SpaceSavingSummaryProfileStatistic(DataProfile refProf) {
+        super(StatisticTitle.hist, StatisticCategory.histCat, refProf);
+    }
+
+    @Override
+    public void calculationNumeric(List<Number> list, Object oldVal) throws NoSuchMethodException {
+
+    }
+
+    @Override
+    public boolean checkConformance(ProfileStatistic m, double threshold) {
+        return false;
+    }
+}
